+++ src/main/java/firok/tiths/common/Alloys.java
@@ -12,7 +12,9 @@ public class Alloys
 	public static void registerAlloys()
 	{
 		TinkersThings.log("register alloys...");
-		registerAlloy(new FluidStack(Fluids.moltenRoyalAlloy, VALUE_Ingot*3),
+		// 皇家合金
+		registerAlloy(
+				new FluidStack(Fluids.moltenRoyalAlloy, VALUE_Ingot*3),
 				new FluidStack(TinkerFluids.iron,VALUE_Ingot*2),
 				new FluidStack(TinkerFluids.gold,VALUE_Ingot)
 				);