+++ src/main/java/firok/tiths/common/TCMaterials.java
@@ -4,8 +4,8 @@ import firok.tiths.util.*;
 import slimeknights.tconstruct.library.materials.Material;
 import slimeknights.tconstruct.tools.TinkerTraits;
 
+import static firok.tiths.common.Keys.*;
 import static firok.tiths.common.Traits.*;
-import static firok.tiths.util.Keys.*;
 import static slimeknights.tconstruct.library.materials.MaterialTypes.HEAD;
 import static slimeknights.tconstruct.tools.TinkerTraits.*;
 
@@ -16,45 +16,45 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material immersedSilver=new Material(nameImmersedSilver, colorImmersedSilver);
+	public static final Material immersedSilver=new Material(nameImmersedSilver, colorImmersedSilver);
 
 
 	@Compo(name=nameMithril,fluid=nameMithril)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material mithril=new Material(nameMithril, colorMithril);
+	public static final Material mithril=new Material(nameMithril, colorMithril);
 
 	@Compo(name=nameAdamantine,fluid=nameAdamantine)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material adamantine=new Material(nameAdamantine, colorAdamantine)
+	public static final Material adamantine=new Material(nameAdamantine, colorAdamantine)
 			.addTrait(starDashing);
 
 	@Compo(name=nameBlackrock,fluid=nameBlackrock)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material blackrock=new Material(nameBlackrock, colorBlackrock);
+	public static final Material blackrock=new Material(nameBlackrock, colorBlackrock);
 
 	@Compo(name=nameInertWitherium,fluid=nameInertWitherium)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material inertWitherium=new Material(nameInertWitherium, colorInertWitherium);
+	public static final Material inertWitherium=new Material(nameInertWitherium, colorInertWitherium);
 
 	@Compo(name=nameWitherium,fluid=nameWitherium)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material witherium=new Material(nameWitherium, colorWitherium);
+	public static final Material witherium=new Material(nameWitherium, colorWitherium);
 
 	@Compo(name=nameRoyalAlloy,fluid=nameRoyalAlloy)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material royalAlloy=new Material(nameRoyalAlloy, colorRoyalAlloy)
+	public static final Material royalAlloy=new Material(nameRoyalAlloy, colorRoyalAlloy)
 			.addTrait(luxurious)
 			.addTrait(TinkerTraits.magnetic2);
 
@@ -62,27 +62,27 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material stellarium=new Material(nameStellarium, colorStellarium)
+	public static final Material stellarium=new Material(nameStellarium, colorStellarium)
 			.addTrait(radiant);
 
 	@Compo(name=nameSpiderLeg)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoExtra(extraDurability = 100)
-	public static Material spiderLeg=new Material(nameSpiderLeg, colorSpiderLeg)
+	public static final Material spiderLeg=new Material(nameSpiderLeg, colorSpiderLeg)
 			.addTrait(TinkerTraits.sharp)
 			.addTrait(TinkerTraits.poisonous);
 
 	@Compo(name=nameHardSpiderLeg)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoExtra(extraDurability = 100)
-	public static Material hardSpiderLeg=new Material(nameHardSpiderLeg, colorHardSpiderLeg)
+	public static final Material hardSpiderLeg=new Material(nameHardSpiderLeg, colorHardSpiderLeg)
 			.addTrait(TinkerTraits.sharp);
 
 	@Compo(name=nameCinnabar,fluid=nameCinnabar)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material cinnabar=new Material(nameCinnabar,colorCinnabar)
+	public static final Material cinnabar=new Material(nameCinnabar,colorCinnabar)
 			.addTrait(TinkerTraits.poisonous, HEAD);
 
 //	@Compo(name= nameGlass, item="glass",fluid="glass")
@@ -97,7 +97,7 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material consolidatedGlass=new Material(nameConsolidatedGlass, colorConsolidatedGlass)
+	public static final Material consolidatedGlass=new Material(nameConsolidatedGlass, colorConsolidatedGlass)
 			.addTrait(cheap)
 			.addTrait(lightweight);
 
@@ -105,7 +105,7 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material brokenIce=new Material(nameBrokenIce,colorBrokenIce)
+	public static final Material brokenIce=new Material(nameBrokenIce,colorBrokenIce)
 			.addTrait(cheap)
 			.addTrait(icy);
 
@@ -113,27 +113,27 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material shell=new Material(nameShell,colorShell)
+	public static final Material shell=new Material(nameShell,colorShell)
 			.addTrait(sharp);
 
 	@Compo(name= nameRuby, item=nameRuby,castatble = false,craftable = true)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material ruby=new Material(nameRuby,colorRuby);
+	public static final Material ruby=new Material(nameRuby,colorRuby);
 
 	@Compo(name= nameCorundum, item=nameCorundum,castatble = false,craftable = true)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material corundum=new Material(nameCorundum,colorCorundum)
+	public static final Material corundum=new Material(nameCorundum,colorCorundum)
 			.addTrait(clustering);
 
 	@Compo(name= nameTitanium, item=nameTitanium,castatble = true,craftable = false)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material titanium=new Material(nameTitanium,colorTitanium)
+	public static final Material titanium=new Material(nameTitanium,colorTitanium)
 			.addTrait(sharp)
 			.addTrait(lightweight);
 
@@ -141,25 +141,49 @@ public class TCMaterials
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material polarium=new Material(namePolarium,colorPolarium);
+	public static final Material polarium=new Material(namePolarium,colorPolarium);
 
 	@Compo(name= nameHalleium, item=nameHalleium,castatble = true,craftable = false)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material halleium=new Material(nameHalleium,colorHalleium);
+	public static final Material halleium=new Material(nameHalleium,colorHalleium);
 
 	@Compo(name= nameAltairium, item=nameAltairium,castatble = true,craftable = false)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material altairium=new Material(nameAltairium,colorAltairium);
+	public static final Material altairium=new Material(nameAltairium,colorAltairium);
 
 	@Compo(name= nameCocoa, item=nameCocoa,castatble = true,craftable = false)
 	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
 	@CompoHandle(modifier = 0.8f, durability = 100)
 	@CompoExtra(extraDurability = 100)
-	public static Material cocoa=new Material(nameCocoa,colorCocoa);
+	public static final Material cocoa=new Material(nameCocoa,colorCocoa);
+
+	@Compo(name= nameNitre, item=nameNitre,castatble = false,craftable = true)
+	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
+	@CompoHandle(modifier = 0.8f, durability = 100)
+	@CompoExtra(extraDurability = 100)
+	public static final Material nitre=new Material(nameNitre,colorNitre)
+			.addTrait(soluble);
+	static
+	{
+		nitre.addItem(Items.nitre);
+	}
+
+	@Compo(name= nameIcelandSpar, item=nameIcelandSpar,castatble = false,craftable = true)
+	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
+	@CompoHandle(modifier = 0.8f, durability = 100)
+	@CompoExtra(extraDurability = 100)
+	public static final Material icelandSpar=new Material(nameIcelandSpar,colorIcelandSpar)
+			.addTrait(birefringent);
+
+	@Compo(name= namePyrophyllite, item=namePyrophyllite,castatble = false,craftable = true)
+	@CompoHead(durability = 100,miningspeed = 0.5f,attack = 1,harvestLevel = 1)
+	@CompoHandle(modifier = 0.8f, durability = 100)
+	@CompoExtra(extraDurability = 100)
+	public static final Material pyrophyllite=new Material(namePyrophyllite,colorPyrophyllite);
 
 	public static Material cloud=new Material("cloud", Colors.Silver);
 	public static Material cloudStorm=new Material("cloud_storm", Colors.DarkBlue);